OkTableTools

Documentation for OkTableTools.

OkTableTools.nearestrows

OkTableTools.sortbydist

OkTableTools.nearestrows — Method

nearestrows returns the row::DataFrameRow that best approaches the given columns and values as pairs. That is, nearestrows(df::AbstractDataFrame, prs::Pair...) returns the row of df where the norm of the distances between column vectors and reference values in pair is minimum.

Example

# to find the `row::DataFrameRow` where the `2`-`norm` of `row.time .- 2000` and  `row.friction .- 0.75` is minimum.
prs = "time" => 2000, "friction" => 0.75

OkTableTools.sortbydist — Method

sortbydist(df::AbstractDataFrame, cols::Cols, vals) sort df by the distances between scalar value and vector variable for value in vals and variable in eachcol(df[!, cols]).